backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
twilio==9.3.7
geopy==2.4.1
python-multipart==0.0.18
